fix(rpc): export types
